barrido y limpieza, refactor de paquetes y otras cositas
