+++ src/main/java/main/modelo/direccion/Direccion.java
@@ -1,5 +1,7 @@
 package main.modelo.direccion;
 
+import main.modelo.tablero.distancia.Posicion;
+
 public abstract class Direccion {
     public abstract Posicion posicionRelativa();
