+++ src/main/java/main/modelo/direccion/Noreste.java
@@ -1,5 +1,7 @@
 package main.modelo.direccion;
 
+import main.modelo.tablero.distancia.Posicion;
+
 public class Noreste extends Direccion{
     public Posicion posicionRelativa(){
         return new Posicion(-1,1);
