+++ src/main/java/main/modelo/direccion/Noroeste.java
@@ -1,5 +1,7 @@
 package main.modelo.direccion;
 
+import main.modelo.tablero.distancia.Posicion;
+
 public class Noroeste extends Direccion{
     public Posicion posicionRelativa(){
         return new Posicion(-1,-1);
