+++ src/main/java/main/modelo/direccion/Norte.java
@@ -1,5 +1,7 @@
 package main.modelo.direccion;
 
+import main.modelo.tablero.distancia.Posicion;
+
 public class Norte extends Direccion{
     public Posicion posicionRelativa(){
         return new Posicion(-1,0);
