+++ src/main/java/main/modelo/direccion/Oeste.java
@@ -1,5 +1,7 @@
 package main.modelo.direccion;
 
+import main.modelo.tablero.distancia.Posicion;
+
 public class Oeste extends Direccion{
     public Posicion posicionRelativa(){
         return new Posicion(0,-1);
