+++ src/main/java/main/modelo/tablero/distancia/Lejana.java
@@ -1,4 +1,4 @@
-package main.modelo.tablero.Distancia;
+package main.modelo.tablero.distancia;
 
 import main.modelo.unidad.Unidad;
