+++ src/main/java/main/modelo/unidad/Soldado.java
@@ -3,6 +3,9 @@ package main.modelo.unidad;
 import main.modelo.direccion.Direccion;
 import main.modelo.unidad.ataque.AtaqueCorto;
 import main.modelo.unidad.conjuntodeunidades.ConjuntoDeSoldados;
+import main.modelo.unidad.movimiento.MovimientoEnBatallon;
+import main.modelo.unidad.movimiento.MovimientoEstrategia;
+import main.modelo.unidad.movimiento.MovimientoRegular;
 
 import java.util.Iterator;
