+++ src/main/java/main/modelo/unidad/Unidad.java
@@ -1,7 +1,7 @@
 package main.modelo.unidad;
 
 import main.modelo.direccion.Direccion;
-import main.modelo.direccion.Posicion;
+import main.modelo.tablero.distancia.Posicion;
 import main.modelo.juego.Equipo;
 import main.modelo.unidad.ataque.AtaqueEstrategia;
 import main.modelo.unidad.conjuntodeunidades.ConjuntoDeSoldados;
