+++ src/main/java/main/modelo/unidad/movimiento/MovimientoEstatico.java
@@ -1,7 +1,8 @@
-package main.modelo.unidad;
+package main.modelo.unidad.movimiento;
 
 import main.modelo.direccion.Direccion;
 import main.modelo.excepciones.UnidadNoPuedeMoverseException;
+import main.modelo.unidad.Unidad;
 
 public class MovimientoEstatico implements MovimientoEstrategia {
     @Override
