+++ src/main/java/main/modelo/unidad/movimiento/MovimientoEstrategia.java
@@ -1,6 +1,7 @@
-package main.modelo.unidad;
+package main.modelo.unidad.movimiento;
 
 import main.modelo.direccion.Direccion;
+import main.modelo.unidad.Unidad;
 
 public interface MovimientoEstrategia {
     void avanzar(Unidad unidad, Direccion direccion);
