+++ src/main/java/main/modelo/unidad/movimiento/MovimientoRegular.java
@@ -1,7 +1,9 @@
-package main.modelo.unidad;
+package main.modelo.unidad.movimiento;
 
 import main.modelo.direccion.Direccion;
 import main.modelo.tablero.Tablero;
+import main.modelo.unidad.Unidad;
+import main.modelo.unidad.movimiento.MovimientoEstrategia;
 
 public class MovimientoRegular implements MovimientoEstrategia {
     @Override
