+++ src/test/java/main/SoldadoTest.java
@@ -9,6 +9,7 @@ import main.modelo.juego.EquipoNegro;
 import main.modelo.juego.Juego;
 import main.modelo.juego.Jugador;
 import main.modelo.tablero.Tablero;
+import main.modelo.tablero.distancia.Posicion;
 import main.modelo.unidad.*;
 import org.junit.Test;
 import org.junit.After;
