+++ src/test/java/main/TableroTest.java
@@ -1,7 +1,7 @@
 package main;
 
 import main.modelo.direccion.Norte;
-import main.modelo.direccion.Posicion;
+import main.modelo.tablero.distancia.Posicion;
 import main.modelo.excepciones.CasilleroEsDeEnemigoException;
 import main.modelo.excepciones.CasilleroOcupadoException;
 import main.modelo.juego.EquipoBlanco;
